+++ demos/createCssBackdropFilterSeclector.js
@@ -1,22 +1,28 @@
 /**
+ * Generates a set of GUI input options for backdrop-filter CSS options.
  *
- * Requires number-with-checkbox extension.
+ * Requires lil-gui's number-with-checkbox and color-with-checkbox extension.
  *
- * @date 2024-07-14
+ * @author  Ikaros Kappler
+ * @date    2024-07-14
+ * @version 1.0.0
  */
 
 (function (_context) {
+  // +---------------------------------------------------------------------------------
+  // | Convert values in [0..1] to [0..100].
+  // +-------------------------------
   var ratio2pct = function (ratio) {
     return "" + ratio * 100.0 + "%";
   };
 
-  var getBackdropFilterString = function (config, onBackdropStringChange) {
-    // console.log("getOpacityPct", ratio2pct(config.opacity));
-    // console.log("config", config);
+  // +---------------------------------------------------------------------------------
+  // | Generate a CSS backdrop-filter string from the given values.
+  // | Consult the docs for possible values:
+  // |    https://developer.mozilla.org/en-US/docs/Web/CSS/backdrop-filter
+  // +-------------------------------
+  var getBackdropFilterString = function (config) {
     var buffer = [];
-    // if (config.isBlurEnabled) {
-    //   buffer.push("invert(" + ratio2pct(config.invert) + ")");
-    // }
     if (config.isOpacityEnabled) {
       buffer.push("opacity(" + ratio2pct(config.opacity) + ")");
     }
@@ -52,45 +58,86 @@
     return buffer.length === 0 ? "none" : buffer.join(" ");
   };
 
+  // +---------------------------------------------------------------------------------
+  // | Fire the callback on changes.
+  // +-------------------------------
   var fireBackdropStringChange = function (config, onBackdropStringChange) {
-    var newBackdropString = getBackdropFilterString(config, onBackdropStringChange);
+    var newBackdropString = getBackdropFilterString(config);
     onBackdropStringChange(newBackdropString, config);
   };
 
-  var createDefaultConfig = function () {
+  // +---------------------------------------------------------------------------------
+  // | Try to get a string value or fallback from an object.
+  // +-------------------------------
+  var _getString = function (config, name, fallback) {
+    return typeof config[name] === "string" ? config[name] : fallback;
+  };
+
+  // +---------------------------------------------------------------------------------
+  // | Try to get a boolean value or fallback from an object.
+  // +-------------------------------
+  var _getBoolean = function (config, name, fallback) {
+    return typeof config[name] === "boolean" ? config[name] : fallback;
+  };
+
+  // +---------------------------------------------------------------------------------
+  // | Try to get a numeric value or fallback from an object.
+  // +-------------------------------
+  var _getNumber = function (config, name, fallback) {
+    return typeof config[name] === "number" ? config[name] : fallback;
+  };
+
+  // +---------------------------------------------------------------------------------
+  // | Initialize the config and try to retrieve initial values passed by the user.
+  // +-------------------------------
+  var createDefaultConfig = function (initialFilterValues) {
     return {
-      opacity: 0.5,
-      isOpacityEnabled: false,
-      invert: 0.8,
-      isInvertEnabled: false,
-      sepia: 0.9,
-      isSepiaEnabled: false,
-      blur: 2, // px
-      isBlurEnabled: false,
-      brightness: 0.6,
-      isBrightnessEnabled: false,
-      contrast: 0.9,
-      isContrastEnabled: false,
-      dropShadow: 4, // px
+      // This is just for the global effect color
+      effectFilterColor: _getString(initialFilterValues, "effectFilterColor", "#204a87"),
+      isEffectsColorEnabled: _getBoolean(initialFilterValues, "isEffectsColorEnabled", true),
+      // These are real filter values
+      opacity: _getNumber(initialFilterValues, "opacity", 0.5),
+      isOpacityEnabled: _getBoolean(initialFilterValues, "isOpacityEnabled", false),
+      invert: _getNumber(initialFilterValues, "invert", 0.8),
+      isInvertEnabled: _getBoolean(initialFilterValues, "isInvertEnabled", false),
+      sepia: _getNumber(initialFilterValues, "sepia", 0.9),
+      isSepiaEnabled: _getBoolean(initialFilterValues, "isSepiaEnabled", false),
+      blur: _getNumber(initialFilterValues, "blur", 2), // px
+      isBlurEnabled: _getBoolean(initialFilterValues, "isBlurEnabled", false),
+      brightness: _getNumber(initialFilterValues, "brightness", 0.6),
+      isBrightnessEnabled: _getBoolean(initialFilterValues, "isBrightnessEnabled", false),
+      contrast: _getNumber(initialFilterValues, "contrast", 0.9),
+      isContrastEnabled: _getBoolean(initialFilterValues, "isContrastEnabled", false),
+      dropShadow: _getNumber(initialFilterValues, "dropShadow", 4), // px
       dropShadowColor: "#00ffff", // HOW TO DISABLE THIS PROPERLY
-      isDropShadowEnabled: false,
-      grayscale: 0.3,
-      isGrayscaleEnabled: false,
-      hueRotate: 120, // deg
-      isHueRotateEnabled: false,
-      saturate: 1.0,
-      isSaturateEnabled: false
+      isDropShadowEnabled: _getBoolean(initialFilterValues, "isDropShadowEnabled", false),
+      grayscale: _getNumber(initialFilterValues, "grayscale", 0.3),
+      isGrayscaleEnabled: _getBoolean(initialFilterValues, "isGrayscaleEnabled", false),
+      hueRotate: _getNumber(initialFilterValues, "hueRotate", 120), // deg
+      isHueRotateEnabled: _getBoolean(initialFilterValues, "isHueRotateEnabled", false),
+      saturate: _getNumber(initialFilterValues, "saturate", 2.0),
+      isSaturateEnabled: _getBoolean(initialFilterValues, "isSaturateEnabled", false)
     };
   };
 
-  _context.createCssBackdropFilterSelector = function (guiFolder, onBackdropStringChange) {
-    var config = createDefaultConfig();
+  // +---------------------------------------------------------------------------------
+  // | This is the actual exported function.
+  // | @param {lil-gui.GUI} guiFolder - The folder to add the controls to.
+  // | @param {function(newBackdropFilterString, config)} onBackdropStringChange - A callback to retrieve changes.
+  // | @param {object} initialFilterValues - [optional] An object with initial backdrop-filter values.
+  // +-------------------------------
+  _context.createCssBackdropFilterSelector = function (guiFolder, onBackdropStringChange, initialFilterValues) {
+    var config = createDefaultConfig(initialFilterValues || {});
 
-    var handleChange = function (newValue, filterEnabled) {
+    var handleChange = function (_newValue, _filterEnabled) {
       //   console.log("New value", newValue, filterEnabled);
       fireBackdropStringChange(config, onBackdropStringChange);
     };
 
+    // prettier-ignore
+    guiFolder
+      .addColorWithCheckbox(config, "effectFilterColor", "isEffectsColorEnabled")
+      .onChange(handleChange);
     // prettier-ignore
     guiFolder
         .addNumberWithCheckbox(config, "opacity", "isOpacityEnabled").onChange(handleChange)
@@ -134,6 +181,4 @@
 
     return handleChange;
   }; // END funcition createCssBackdropFilterSelector
-
-  console.log("createCssBackdropFilterSelector", createCssBackdropFilterSelector);
 })(globalThis);