+++ changelog.md
@@ -17,6 +17,8 @@ Todo:
 - replace all colors + alphas by the new input extension (lil-gui).
 - Add googley eyes to the animated metaballs demo.
 
+* 2024-07-23
+  - Finalizing the CSS backdrop-filter effects demo (54).
 * 2024-07-17
   - Tweaking the lil-gui extension to handle different input types with checkboxes.
 * 2024-07-15